+++ src/components/app/routes/data/utils.js
@@ -21,43 +21,15 @@ import {
   queryContentHighlightsConfiguration,
   queryCouponCodeRequests,
   queryCouponCodes,
-  queryEnterpriseLearnerDashboardBFF,
   queryEnterpriseLearnerOffers,
-  queryFallbackRouteBFF,
   queryLicenseRequests,
   queryNotices,
   queryRedeemablePolicies,
   querySubscriptions,
+  resolveBFFQuery,
   transformSubscriptionsData,
   updateUserActiveEnterprise,
 } from '../../data';
-
-/**
- * Resolves the appropriate BFF query function to use for the current route.
- * @param {string} pathname - The current route pathname.
- * @returns {Function|null} The BFF query function to use for the current route, or null if no match is found.
- */
-export function resolveBFFQuery(pathname) {
-  // Define route patterns and their corresponding query functions
-
-  const routeToBFFQueryMap = [
-    {
-      pattern: '/:enterpriseSlug',
-      query: queryEnterpriseLearnerDashboardBFF,
-    },
-    // Add more routes and queries incrementally as needed
-  ];
-
-  // Find the matching route and return the corresponding query function
-  const matchedRoute = routeToBFFQueryMap.find((route) => matchPath(route.pattern, pathname));
-
-  if (matchedRoute) {
-    return matchedRoute.query;
-  }
-
-  // No match found
-  return queryFallbackRouteBFF;
-}
 
 /**
  * Ensures all enterprise-related app data is loaded.
@@ -82,9 +54,9 @@ export async function ensureEnterpriseAppData({
 }) {
   const shouldUseBFF = isBFFFeatureFlagEnabled(enterpriseCustomer.uuid);
   const matchedBFFQuery = resolveBFFQuery(requestUrl.pathname);
-  const subscriptionsQuery = shouldUseBFF || !matchedBFFQuery(params).queryKey.includes('fallback')
-    ? matchedBFFQuery(params)
-    : querySubscriptions(enterpriseCustomer.uuid);
+  const subscriptionsQuery = !shouldUseBFF || matchedBFFQuery(params).queryKey.includes('fallback')
+    ? querySubscriptions(enterpriseCustomer.uuid)
+    : matchedBFFQuery(params);
   const enterpriseAppDataQueries = [
     // Enterprise Customer User Subsidies
     queryClient.ensureQueryData(subscriptionsQuery).then(async (rootLoaderData) => {
