+++ src/index.jsx
@@ -55,14 +55,9 @@ initialize({
         FEATURE_CONTENT_HIGHLIGHTS: process.env.FEATURE_CONTENT_HIGHLIGHTS || null,
         FEATURE_ENABLE_EMET_REDEMPTION: process.env.FEATURE_ENABLE_EMET_REDEMPTION || null,
         ENTERPRISE_SUBSIDY_BASE_URL: process.env.ENTERPRISE_SUBSIDY_BASE_URL || null,
-        // TODO: REMOVE ME
-        // Disabled
-        FEATURE_ENABLE_BFF_API_FOR_ENTERPRISE_CUSTOMERS: [] || null,
-        // Enabled
-        // FEATURE_ENABLE_BFF_API_FOR_ENTERPRISE_CUSTOMERS: ['d0224477-4903-404a-b3c4-9edfce728d37'] || [],
-        // FEATURE_ENABLE_BFF_API_FOR_ENTERPRISE_CUSTOMERS: (
-        // process.env.FEATURE_ENABLE_BFF_API_FOR_ENTERPRISE_CUSTOMERS || []
-        // ),
+        FEATURE_ENABLE_BFF_API_FOR_ENTERPRISE_CUSTOMERS: (
+          process.env.FEATURE_ENABLE_BFF_API_FOR_ENTERPRISE_CUSTOMERS || []
+        ),
         EXPERIMENT_2_ID: process.env.EXPERIMENT_2_ID || null,
         EXPERIMENT_2_VARIANT_2_ID: process.env.EXPERIMENT_2_VARIANT_2_ID || null,
         FEATURE_ENABLE_RESTRICTED_RUNS: process.env.FEATURE_ENABLE_RESTRICTED_RUNS || null,
